added the intern view for the host profile
